refactor: Separate Pik class into separate mdoule
