+++ .vscode/settings.json
@@ -8,7 +8,8 @@
         "ModCallbacks",
         "Vector",
         "NpcState",
-        "EntityCollisionClass"
+        "EntityCollisionClass",
+        "FamiliarVariant"
     ],
     "editor.wordWrap": "on"
 }
